move files from bin/ to tool/
