+++ dart/pubspec.yaml
@@ -1,14 +1,17 @@
 name: vm_service_lib
 description: A library to access the VM Service API.
 version: 0.1.0
-#author: Your Name <email@example.com>
-#homepage: https://www.example.com
+
+author: Dart Team <misc@dartlang.org>
+homepage: https://github.com/dart-lang/vm_service_drivers
 
 environment:
   sdk: '>=1.8.0 <2.0.0'
 
 dependencies:
-  markdown: ^0.7.0
+  logging: any
 
 dev_dependencies:
+  markdown: ^0.7.0
+  path: ^1.0.0
   test: ^0.12.0
