+++ dart/tool/common/src_gen_common.dart
